Refactor config loading code
