+++ hw.js
@@ -41,57 +41,9 @@ if(!command) {
     process.exit(0);
 }
 
-// import configuration file conditionally
+var config = require("./current-config.js");
 
-var config;
-
-try {
-    config = require(process.cwd() + "/config.js");
-    main();
-} catch(e) {
-    // if this is init command, that's fine :)
-    
-    if(command != "init") {
-        // alright, perhaps there is a default hw instance somewhere else,
-        // and we can change dir there instead:
-        
-        function failMessage() {
-            console.error("The configuration file could not be loaded.");
-            console.error("Did you use `hw init` first?");
-            console.error("If so, this is a bug. https://github.com/bobbybee/hw/issues/new");
-            process.exit(1);
-        }
-
-        console.log("Searching "+process.env["HOME"]+"/.hw_default");
-
-        fs.readFile(process.env["HOME"]+"/.hw_default", function(err, data) {
-            if(err) {
-                throw err;
-                failMessage();
-            }
-            
-
-            var n = data.toString().trim();
-
-            // alright, let's change directories and try again
-            try {
-                process.chdir(n);
-            } catch(e) {
-                console.error(e);
-                console.warn(process.cwd());
-                failMessage();
-            }
-
-            try {
-                config = require(process.cwd() + "/config.js");
-                main();
-            } catch(e) {
-                // our options are exhausted at this point :(
-                failMessage();
-            }
-        });
-    }
-}
+main();
 
 function main() {
     if(command == "add") {
